- Implemented path loading/plotting for alexi-sim  - Updated trainer/RNN parameters (functions, learning rate, bptt steps, etc)  - Minor fixes

+++ libaann/include/training_set.hpp
@@ -10,7 +10,7 @@ struct frame_data{
     double rw, rl;
 
     double x, y;
-    double px, py;
+    int px, py;
     double dx, dy;
     double dx_last, dy_last;
     double theta;
